reorganize_folder

+++ shapmagn/shape/shape_utils.py
@@ -10,4 +10,4 @@ def get_scale_and_center(points,percentile=99):
         filtered_up_thre =  np.percentile(points[:,d],interval[1])
         scale[0,d] = filtered_up_thre - filtered_low_thre
         center[0,d] = (filtered_up_thre + filtered_low_thre)/2
-    return scale/2, center
+    return (scale/2).astype(np.float32), center.astype(np.float32)
